backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
frozenlist==1.4.1
idna==3.7
ipinfo==4.2.1
ipinfo-django==2.0.0
multidict==6.0.5
requests==2.31.0
sqlparse==0.5.0
urllib3==2.2.1
yarl==1.9.4
